move countries table on separate page
