+++ src/pages/Countries/countries.scss
@@ -1,0 +1,5 @@
+.countries-page {
+  &__search-bar {
+    width: 100%;
+  }
+}
