--- src/pages/Home/home.scss
@@ -1,5 +1,0 @@
-.home-page {
-  &__search-bar {
-    width: 100%;
-  }
-}
