+++ src/router/Router.tsx
@@ -1,9 +1,11 @@
 import React, { Suspense } from "react";
-import { BrowserRouter, Redirect, Route, Switch } from "react-router-dom";
+import { BrowserRouter, Route, Switch } from "react-router-dom";
 
 const Header = React.lazy(() => import("../components/Header"));
 const Footer = React.lazy(() => import("../components/Footer"));
 const Home = React.lazy(() => import("../pages/Home"));
+const Countries = React.lazy(() => import("../pages/Countries"));
+const Country = React.lazy(() => import("../pages/Country"));
 
 const Router = () => {
   return (
@@ -11,6 +13,9 @@ const Router = () => {
       <BrowserRouter>
         <Header />
         <Switch>
+          <Route path="/countries">
+            <Countries />
+          </Route>
           <Route path="/">
             <Home />
           </Route>
